--- src/main/java/com/newbit/coffeechat/command/domain/repository/ProductRepository.java
@@ -1,4 +1,0 @@
-package com.newbit.coffeechat.command.domain.repository;
-
-public interface ProductRepository {
-}
