+++ src/main/java/com/newbit/coffeechat/command/infrastructure/repository/JpaCoffeechatRepository.java
@@ -1,6 +1,6 @@
 package com.newbit.coffeechat.command.infrastructure.repository;
 
-import com.newbit.coffeechat.command.application.domain.repository.CoffeechatRepository;
+import com.newbit.coffeechat.command.domain.repository.CoffeechatRepository;
 import com.newbit.coffeechat.command.domain.aggregate.Coffeechat;
 import org.springframework.data.jpa.repository.JpaRepository;
